Deleted some files and added it to other files
